+++ CHANGELOG.rst
@@ -1,7 +1,7 @@
 Changelog
 =========
 
-Release *v2.0.dev0* - ``2019-10-07``
+Release *v2.0.dev9* - ``2019-10-08``
 ------------------------------------
 * Support for redis-py >= 3 only
 * Support for redis-server >= 3 only
@@ -9,6 +9,17 @@ Release *v2.0.dev0* - ``2019-10-07``
 * Breaking change: `zadd` flags other than `ch` are explicitely not supported
 * Breaking change: `zincrby` arguments are swaped (`amount, value` instead of `value, amount`)
 * Breaking change: Redis server with `LUA` scripting support is mandatory
+* Breaking change: `set` flags other than `ex` and `px` are explicitely not supported
+* Add `decrby`, `incrby` and `bitpos` to `StringField`
+* Add expiring commands to all normal fields (not `InstanceHashField` and `*PKField`): `expire`, `pexpire`, `expireat`, `pexpireat`, `ttl`, `pttl`, `persite`. But `*expire*` commands can only be called on non-indexable fields
+* Add `setex` and `psetex` to `StringField`. Can only be called on non-indexable fields.
+* Deny `ex` and `px` flag to `set` if field is indexable
+* Add support for `count` argument to `spop` (only for redis-server >= 3.2)
+* Add new commands to `ListField`: `lcontains`, `lrank` and `lcount`, to know if a value is in the list, where, and how many times. This is done on the redis server side via lua scripting.
+* Optimize deindexing when calling `hmset` or `hdel`
+* Add `hstrlen` to `HashField`
+* Add `zlexcount`, `zrangebylex`, `zremrangebylex` and `zrevrangebylex` to `SortedSetField`
+* Add `zpopmax` and `zpopmin` `SortedSetField` (only for redis-server >= 5)
 
 Release *v1.3.1* - ``2019-10-11``
 ---------------------------------